+++ 2020/Snake/food.js
