+++ 2020/Snake/input.js
