--- game.js
@@ -1,49 +1,0 @@
-import { update as updateSnake, draw as drawSnake, snakeSpeed, getSnakeHead, snakeIntersection } from './snake.js'
-import { update as updateFood, draw as drawFood } from './food.js'
-import { outsideGrid } from './grid.js'
-
-
-let lastRenderTime = 0
-let gameOver = false
-const gameBoard = document.getElementById("game-board")
-
-
-function gameSpeed(currentTime) {
-    if (gameOver) {
-        if (confirm('You lost. Press ok to restart.')) {
-            window.location = '/'
-        }
-        return
-    }
-
-    window.requestAnimationFrame(gameSpeed)
-    const secondsSinceLastRender = (currentTime - lastRenderTime) / 1000
-    if (secondsSinceLastRender < 1 / snakeSpeed) return
-    
-    lastRenderTime = currentTime
-    
-    update()
-    draw()
-}
-
-
-window.requestAnimationFrame(gameSpeed)
-
-
-function update() {
-    updateSnake()
-    updateFood()
-    checkDeath()
-}
-
-
-function draw() {
-    gameBoard.innerHTML = ''
-    drawSnake(gameBoard)
-    drawFood(gameBoard)
-}
-
-
-function checkDeath() {
-    gameOver = outsideGrid(getSnakeHead()) || snakeIntersection()
-}
