--- snake.html
@@ -1,60 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=<device-width>, initial-scale=1.0">
-
-    <title>SnakeGame</title>
-
-
-    <style>
-
-
-        body {
-            height:100vh;
-            width: 100vw;
-            display: flex;
-            justify-content: center;
-            align-items: center;
-            margin: 0;
-            background-color: black;
-        }
-
-        #game-board{
-            background-color: hsl(0, 0%, 25%);
-            width: 100vmin;
-            height: 100vmin;
-            display: grid;
-            grid-template-rows: repeat(21, 1fr);
-            grid-template-columns: repeat(21, 1fr);
-        }
-
-        .snake{
-            background-color: hsl(135, 55%, 44%);
-            border: 0.25vmin solid black;
-        }
-
-        .food{
-            background-color: hsl(50, 100%, 50%);
-            border: 0.25vmin solid black;
-
-        }
-
-
-    </style>
-
-
-    <script src="game.js" defer type= "module"></script>
-
-
-</head>
-
-
-<body>
-
-
-    <div id="game-board"></div>
-
-
-</body>
-</html>
